Add testing for extracting additional content
