+++ pixaxe/pixaxe.py
@@ -1,12 +1,10 @@
 import hashlib
 import os
 import re
-import struct
 import subprocess
 from tempfile import NamedTemporaryFile
 from typing import Optional
 
-import magic
 from assemblyline.common.str_utils import safe_str
 from assemblyline.odm.base import FULL_URI
 from assemblyline_v4_service.common.base import ServiceBase
@@ -26,167 +24,18 @@ from PIL import ImageFile, UnidentifiedImageError
 from stegano import lsb
 from wand.image import Image
 
-from .steg import ImageInfo, NotSupported
+from pixaxe.steg import ImageInfo, NotSupported
+from pixaxe.helper import find_additional_content
 
 ImageFile.LOAD_TRUNCATED_IMAGES = True
 
 
 class Pixaxe(ServiceBase):
-    PAT_FILEMARKERS = {
-        # Header, Trailer, additional methods
-        "bmp": ["\x42\x4D", None, "bmp_dump"],
-        "gif": ["\x47\x49\x46\x38.\x61.{19,}\x2C.{9,}", "\x00\x3B", None],
-        "jpeg": ["\xFF\xD8.{18}\xFF\xDB.{3,}\xFF\xDA.{13,}", "\xFF\xD9", None],
-        "jpeg2000": ["\x00\x00\x00\x0C\x6A\x50\x20\x20\x0D\x0A", None, "jp2_dump"],
-        "png": ["\x89\x50\x4E\x47", "\x49\x45\x4E\x44.{4}", None],
-    }
-
     def __init__(self, config=None):
         super(Pixaxe, self).__init__(config)
 
     def start(self):
         self.log.debug("Pixaxe service started")
-
-    @staticmethod
-    def mimetype(f, t):
-        """Determine if Magic-MIME file type of data matches desired file type.
-
-        Args:
-            f: Raw data to evaluate.
-            t: File type to compare (string).
-
-        Returns:
-            True if file type matches t, or False.
-        """
-        is_t = False
-        m = magic.Magic(mime=True)
-        ftype = m.from_buffer(f)
-        if t in ftype:
-            is_t = True
-        return is_t
-
-    def bmp_dump(self, data):
-        """BMP embedded file extraction. Looks for traits of known BMP file structure to find embedded BMP data.
-
-        Args:
-            data: Raw data to search.
-
-        Returns:
-            BMP data if discovered, or original data.
-        """
-        # noinspection PyBroadException
-        try:
-            # Byte offset to start of image
-            soi = struct.unpack("<I", data[10:14])[0]
-            # Size of image data, including padding -- potentially unreliable
-            sizei = struct.unpack("<I", data[34:38])[0]
-            # Width in pixels
-            # wi = struct.unpack('<I', data[18:22])[0]
-            # Height in pixels
-            # hi = struct.unpack('<I', data[22:26])[0]
-            # Depth
-            # di = struct.unpack('<H', data[26:28])[0]
-            # Bits per pixel
-            # bpi = struct.unpack('<H', data[28:30])[0]
-            # Image bytes
-            # sizei = (wi*hi*di*bpi)/8
-            bmp_data = data[0 : (soi + sizei)]
-            verify_bmp = self.mimetype(bmp_data, "image")
-            if not verify_bmp:
-                return data
-            return bmp_data
-        except Exception:
-            return data
-
-    def jpg2_dump(self, data):
-        """Looks for traits of known JPEG 2000 file structure to confirm that data is likely JPEG 2000 data.
-
-        Args:
-            data: Raw data to search.
-
-        Returns:
-            JPEG 2000 data if discovered, or original data.
-        """
-        ftyps = {
-            "\x6a\x70\x32\x20": "jp2",
-            "\x6a\x70\x78\x20": "jpf",
-            "\x6a\x70\x6d\x20": "jpm",
-            "\x6d\x6a\x70\x32": "mj2",
-            "\xFF\x4F\xFF\x51": "j2c",
-        }
-        trailer = "\xFF\xD9"
-        cdata = data
-        end = 0
-        try:
-            jtype = data[20:24]
-            if jtype in ftyps:
-                file_type = ftyps[jtype]
-                self.log.debug(file_type)
-            else:
-                return
-            while True:
-                findend = cdata.find(trailer)
-                if findend == -1:
-                    return
-                else:
-                    end += findend + 2
-                # Another jp2 codestream
-                if cdata[findend + 6 : findend + 10] == "jp2c":
-                    cdata = cdata[findend + 2 :]
-                # Possible .mov file types
-                elif file_type == "mj2" and cdata[findend + 6 : findend + 10] in [
-                    "free",
-                    "mdat",
-                    "moov",
-                    "pnot",
-                    "skip",
-                    "wide",
-                ]:
-                    msize = struct.unpack(">I", cdata[findend + 2 : findend + 6])[0]
-                    jp2_data = data[0 : end + msize]
-                    break
-                else:
-                    jp2_data = data[0:end]
-                    break
-            return jp2_data
-        except Exception:
-            return
-
-    def find_additional_content(self, alfile):
-        """Looks for appended file content attached to an image.
-
-        Args:
-            alfile: AL submission file path.
-
-        Returns:
-            Embedded file data if found, or None.
-        """
-        with open(alfile, "rb") as f:
-            data = f.read()
-        for _, tinfo in iter(self.PAT_FILEMARKERS.items()):
-            # Build up the regex
-            if tinfo[1] is not None:
-                embed_regex = re.compile(tinfo[0] + ".+" + tinfo[1], re.DOTALL)
-            else:
-                embed_regex = re.compile(tinfo[0] + ".+", re.DOTALL)
-
-            # Find the pattern that should match the image.
-            img_match = re.match(embed_regex, str(data))
-            if img_match:
-                img_data = img_match.group()
-                # Go to extraction module if there is one
-                if tinfo[2] is not None:
-                    img_data = getattr(self, tinfo[2])(img_data)
-
-                # Otherwise extract data as-is (regex is considered good enough)
-                leftovers = data.replace(img_data, b"")
-
-                # Remove trailing NULL bytes
-                leftovers = re.sub(b"[\x00]*$", b"", leftovers)
-
-                if len(leftovers) > 15:
-                    return leftovers
-        return
 
     def tag_network_iocs(self, section: ResultSection, ocr_io: NamedTemporaryFile) -> None:
         ocr_io.seek(0)
@@ -277,7 +126,6 @@ class Pixaxe(ServiceBase):
     def execute(self, request: ServiceRequest):
         """Main Module. See README for details."""
         result = Result()
-
         displayable_image_path = request.file_path
         pillow_incompatible = False
 
@@ -431,7 +279,7 @@ class Pixaxe(ServiceBase):
         # Default to original behaviour
         else:
             # Find attached data
-            additional_content = self.find_additional_content(request.file_path)
+            additional_content = find_additional_content(request.file_contents)
             if additional_content:
                 ares = ResultMemoryDumpSection("Possible Appended Content Found")
                 ares.add_line("{} Bytes of content found at end of image file".format(len(additional_content)))
@@ -441,11 +289,11 @@ class Pixaxe(ServiceBase):
                 result.add_section(ares)
                 file_name = "{}_appended_img_content".format(hashlib.sha256(additional_content).hexdigest()[0:10])
                 file_path = os.path.join(self.working_directory, file_name)
+                with open(file_path, "wb") as unibu_file:
+                    unibu_file.write(additional_content)
                 request.add_extracted(
                     file_path, file_name, "Carved content found at end of image.", safelist_interface=self.api_interface
                 )
-                with open(file_path, "wb") as unibu_file:
-                    unibu_file.write(additional_content)
 
         # Steganography modules
         try:
