+++ tests/test_pixaxe_samples.py
@@ -3,6 +3,7 @@ import os
 import pytest
 from assemblyline.common.importing import load_module_by_path
 from assemblyline_service_utilities.testing.helper import TestHelper
+from pixaxe.helper import find_additional_content
 
 # Force manifest location
 os.environ["SERVICE_MANIFEST_PATH"] = os.path.join(os.path.dirname(__file__), "..", "service_manifest.yml")
@@ -19,3 +20,11 @@ th = TestHelper(service_class, RESULTS_FOLDER, SAMPLES_FOLDER)
 @pytest.mark.parametrize("sample", th.result_list())
 def test_sample(sample):
     th.run_test_comparison(sample)
+
+
+@pytest.mark.parametrize("sample", ["complex.jpg", "simple.gif", "helloworld.bmp"])
+def test_find_additional_content(sample):
+    with open(os.path.join(SAMPLES_FOLDER, sample), "rb") as fp:
+        data = fp.read()
+
+    assert find_additional_content(data)
